* upstream/master-2.4: (4245 commits)
  debian: Update symbols files
  debian: Update packaging to match new SO versions
  tshark: fix pdml export for LUA cases - bug fix for wrong level introduced in SHA-1:b8ead953 - fix indent for Close off fake wrapper protocol
  [Automatic update for 2017-12-03]
  2.4.3 → 2.4.4.
  Build 2.4.3.
  CIP Safety: base packet length can be used
  NetBIOS: Don't write past the beginning of a buffer.
  CIP Safety: check packet length before dissecting
  Use a separate Boolean to indicate whether we have a duration.
  wireless-timeline: handle generators that report incorrect MCS for some frames
  ieee80211-radio: allow 0 duration and handle missing phy type
  wireless_timeline: fix blank display
  Update the release notes for 2.4.3.
  Report the actual *error* for CANT_GET_INTERFACE_LIST.
  ui: Sort profile names
  gitignore: Add packaging/macosx/PkgInfo
  TLS13: update for draft -22 (Server Hello, HRR, CCS, NST)
  Bugfix DCE/RPC Decode As for GTK.
  *Always* set pinfo->p2p_dir from the wtap_pkthdr.
  ...

Conflicts:
	.gitreview
	.tx/config
	AUTHORS
	CMakeLists.txt
	ChangeLog
	Makefile.am
	NEWS
	README.cmake
	capinfos.c
	captype.c
	caputils/capture-pcap-util.c
	cmake/modules/FindWSWinLibs.cmake
	codecs/CMakeLists.txt
	codecs/Makefile.am
	configure.ac
	debian/changelog
	debian/control
	debian/libwireshark0.install
	debian/libwireshark0.symbols
	debian/libwireshark10.install
	debian/libwireshark10.symbols
	debian/libwireshark8.install
	debian/libwireshark8.symbols
	debian/libwiretap0.docs
	debian/libwiretap0.install
	debian/libwiretap0.symbols
	debian/libwiretap6.docs
	debian/libwiretap6.install
	debian/libwiretap6.symbols
	debian/libwiretap7.docs
	debian/libwiretap7.install
	debian/libwiretap7.symbols
	debian/libwsutil0.install
	debian/libwsutil0.symbols
	debian/libwsutil7.install
	debian/libwsutil7.symbols
	debian/libwsutil8.install
	debian/libwsutil8.symbols
	debian/patches/05_note-README-when-running-as-root.patch
	debian/patches/06_release-version.patch
	debian/patches/16_licence_about_location.patch
	doc/README.plugins
	doc/extcap_example.py
	docbook/README.txt
	docbook/asciidoc.conf
	docbook/release-notes.asciidoc
	docbook/wsdg_src/WSDG_chapter_quick_setup.asciidoc
	dumpcap.c
	editcap.c
	epan/CMakeLists.txt
	epan/Makefile.am
	epan/addr_resolv.c
	epan/color_filters.c
	epan/crypt/airpdcap_debug.h
	epan/dissectors/CMakeLists.txt
	epan/dissectors/Makefile.am
	epan/dissectors/asn1/dop/packet-dop-template.c
	epan/dissectors/asn1/h225/packet-h225-template.c
	epan/dissectors/asn1/h245/packet-h245-template.c
	epan/dissectors/asn1/qsig/packet-qsig-template.c
	epan/dissectors/asn1/ros/packet-ros-template.c
	epan/dissectors/asn1/rtse/packet-rtse-template.c
	epan/dissectors/asn1/s1ap/s1ap.cnf
	epan/dissectors/asn1/spnego/packet-spnego-template.c
	epan/dissectors/packet-amqp.c
	epan/dissectors/packet-atm.c
	epan/dissectors/packet-ber.c
	epan/dissectors/packet-bgp.c
	epan/dissectors/packet-bootp.c
	epan/dissectors/packet-cip.c
	epan/dissectors/packet-cisco-erspan.c
	epan/dissectors/packet-coap.c
	epan/dissectors/packet-dcerpc-eventlog.c
	epan/dissectors/packet-dcerpc-witness.c
	epan/dissectors/packet-dcerpc.c
	epan/dissectors/packet-dcm.c
	epan/dissectors/packet-dcp-etsi.c
	epan/dissectors/packet-diameter.c
	epan/dissectors/packet-dns.c
	epan/dissectors/packet-dof.c
	epan/dissectors/packet-dop.c
	epan/dissectors/packet-dtls.c
	epan/dissectors/packet-dvbci.c
	epan/dissectors/packet-e212.c
	epan/dissectors/packet-eap.c
	epan/dissectors/packet-eigrp.c
	epan/dissectors/packet-enc.c
	epan/dissectors/packet-enip.c
	epan/dissectors/packet-epl.c
	epan/dissectors/packet-erf.c
	epan/dissectors/packet-flip.c
	epan/dissectors/packet-gsm_map.c
	epan/dissectors/packet-gtp.c
	epan/dissectors/packet-gtpv2.c
	epan/dissectors/packet-h225.c
	epan/dissectors/packet-h245.c
	epan/dissectors/packet-h248.c
	epan/dissectors/packet-i2c.c
	epan/dissectors/packet-ieee1722.c
	epan/dissectors/packet-ieee80211-radiotap.c
	epan/dissectors/packet-ieee80211.c
	epan/dissectors/packet-ip.c
	epan/dissectors/packet-ipp.c
	epan/dissectors/packet-ipv6.c
	epan/dissectors/packet-isakmp.c
	epan/dissectors/packet-kdsp.c
	epan/dissectors/packet-kerberos.c
	epan/dissectors/packet-kerberos.h
	epan/dissectors/packet-mac-lte.c
	epan/dissectors/packet-mip6.c
	epan/dissectors/packet-mswsp.c
	epan/dissectors/packet-nas_eps.c
	epan/dissectors/packet-null.c
	epan/dissectors/packet-opensafety.c
	epan/dissectors/packet-per.c
	epan/dissectors/packet-qsig.c
	epan/dissectors/packet-quic.c
	epan/dissectors/packet-raknet.c
	epan/dissectors/packet-ros.c
	epan/dissectors/packet-rpcrdma.c
	epan/dissectors/packet-rtps.c
	epan/dissectors/packet-rtse.c
	epan/dissectors/packet-s1ap.c
	epan/dissectors/packet-sdp.c
	epan/dissectors/packet-selfm.c
	epan/dissectors/packet-socketcan.c
	epan/dissectors/packet-spnego.c
	epan/dissectors/packet-ssl-utils.c
	epan/dissectors/packet-ssl-utils.h
	epan/dissectors/packet-ssl.c
	epan/dissectors/packet-tcp.c
	epan/dissectors/packet-uaudp.c
	epan/dissectors/packet-uftp.c
	epan/dissectors/packet-uftp4.c
	epan/dissectors/packet-umts_fp.c
	epan/dissectors/packet-wbxml.c
	epan/dissectors/packet-wsp.c
	epan/dissectors/packet-xml.c
	epan/dissectors/packet-xml.h
	epan/dissectors/packet-zbee-nwk-gp.c
	epan/dissectors/packet-zvt.c
	epan/dissectors/usb.c
	epan/enterprise-numbers
	epan/follow.c
	epan/packet.c
	epan/print.c
	epan/proto.c
	epan/sminmpec.c
	epan/uat.c
	epan/wslua/make-init-lua.pl
	extcap.c
	extcap.h
	extcap/androiddump.c
	extcap/sshdump.c
	extcap_spawn.c
	macosx-setup.sh
	make-version.pl
	manuf
	manuf.tmpl
	mergecap.c
	packaging/macosx/Wireshark_package.pmdoc/index.xml.in
	packaging/nsis/CMakeLists.txt
	packaging/nsis/config.nsh.in
	packaging/rpm/SPECS/wireshark.spec.in
	packaging/wix/CMakeLists.txt
	packaging/wix/Makefile.am
	plugins/docsis/packet-macmgmt.c
	plugins/docsis/packet-ucd.c
	plugins/ethercat/packet-ethercat-frame.c
	plugins/mate/mate_parser.l
	plugins/profinet/packet-dcerpc-pn-io.c
	randpkt.c
	rawshark.c
	reordercap.c
	ringbuffer.c
	services
	test/suite-decryption.sh
	text2pcap.c
	tfshark.c
	tools/fuzz-test.sh
	tools/make-manuf
	tools/pre-commit
	tools/randpkt-test.sh
	tools/win-setup.ps1
	tshark.c
	ui/export_object_smb.c
	ui/export_object_ui.c
	ui/export_object_ui.h
	ui/gtk/decode_as_dlg.c
	ui/gtk/dfilter_expr_dlg.c
	ui/gtk/export_object_dlg.c
	ui/iface_lists.c
	ui/qt/conversation_dialog.cpp
	ui/qt/export_object_dialog.cpp
	ui/qt/extcap_argument.cpp
	ui/qt/extcap_options_dialog.cpp
	ui/qt/file_set_dialog.cpp
	ui/qt/main_window_slots.cpp
	ui/qt/proto_tree.cpp
	ui/qt/simple_dialog.cpp
	ui/qt/wireshark_de.ts
	ui/qt/wireshark_en.ts
	ui/qt/wireshark_fr.ts
	ui/qt/wireshark_it.ts
	ui/qt/wireshark_ja_JP.ts
	ui/qt/wireshark_pl.ts
	ui/qt/wireshark_zh_CN.ts
	ui/simple_dialog.h
	ui/software_update.c
	ui/text_import.c
	version.conf
	wireshark-qt.cpp
	wiretap/CMakeLists.txt
	wiretap/Makefile.am
	wiretap/erf.c
	wiretap/erf.h
	wiretap/merge.c
	wiretap/wtap.c
	wsutil/CMakeLists.txt
	wsutil/Makefile.am
	wsutil/tempfile.c

+++ epan/dissectors/packet-smb2.h
@@ -81,7 +81,7 @@ typedef struct _smb2_saved_info_t {
 typedef struct _smb2_tid_info_t {
 	guint32 tid;
 	guint32 connect_frame;
-	guint16 share_type;
+	guint8 share_type;
 	char *name;
 } smb2_tid_info_t;
 
@@ -182,7 +182,7 @@ typedef struct _smb2_info_t {
 	guint32 status;
 	guint32 tid;
 	guint64 sesid;
-	gint64  msg_id;
+	guint64  msg_id;
 	guint32 flags;
 	smb2_eo_file_info_t	*eo_file_info; /* eo_smb extra info */
 	smb2_conv_info_t	*conv;
